Rearranged dates information into its own class, Nights
